feat: gate validateFn support behind a build tag
